requirements.txt
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
Pillow==7.1.2
Werkzeug==1.0.1
WTForms==2.3.1
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.27.2
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.9
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
markovify==0.8.0
mccabe==0.6.1
msgpack==1.0.0
opencv-python==4.2.0.32
protobuf==3.11.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.4.4
pynvim==0.4.1
python-telegram-bot==12.4.2
PyYAML==5.3
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
termcolor==1.1.0
tornado==6.0.4
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
wrapt==1.11.2
uwsgi